+++ packages/cmsload/src/factory.ts
@@ -1,6 +1,7 @@
-import type { CMSList } from '@finsweet/attributes-cmscore';
-import { addItemsAnimation, addListAnimation } from '@finsweet/attributes-cmscore';
 import { COMMERCE_CSS_CLASSES, LIGHTBOX_CSS_CLASSES } from '@finsweet/ts-utils';
+
+import { checkCMSCoreVersion, CMSList } from '$packages/cmscore';
+import { addItemsAnimation, addListAnimation } from '$packages/cmscore';
 
 import { initDefaultMode } from './modes/default';
 import { initInfiniteMode } from './modes/infinite';
@@ -71,10 +72,18 @@ export const initLoadInstance = async (listInstance: CMSList) => {
     if (itemsCount) listInstance.addItemsCount(itemsCount);
   }
 
-  // Get page count element
-  if (!listInstance.visibleCount) {
-    const visibleCount = queryElement<HTMLElement>('visibleCount', { instanceIndex });
-    if (visibleCount) listInstance.addVisibleCount(visibleCount);
+  // Get visible count elements
+  if (!listInstance.visibleCount || !listInstance.visibleCountFrom || !listInstance.visibleCountTo) {
+    const visibleCountTotal = queryElement<HTMLElement>('visibleCount', { instanceIndex });
+    const visibleCountFrom = queryElement<HTMLElement>('visibleCountFrom', { instanceIndex });
+    const visibleCountTo = queryElement<HTMLElement>('visibleCountTo', { instanceIndex });
+
+    // TODO: Remove this check after cmscore v1.7.0 has rolled out
+    if (checkCMSCoreVersion('>=', '1.7.0')) {
+      listInstance.addVisibleCount(visibleCountTotal, visibleCountFrom, visibleCountTo);
+    } else if (visibleCountTotal) {
+      listInstance.addVisibleCount(visibleCountTotal);
+    }
   }
 
   // Get scroll anchor