+++ packages/cmssort/src/actions/sort.ts
@@ -1,4 +1,6 @@
-import type { CMSList } from '@finsweet/attributes-cmscore';
+import { normalizeNumber } from 'global/helpers';
+
+import { checkCMSCoreVersion, CMSList } from '$packages/cmscore';
 
 import type { SortingDirection } from '../utils/types';
 
@@ -24,9 +26,9 @@ export const sortListItems = async (
     addingItems?: boolean;
   }
 ) => {
-  const { items } = listInstance;
+  const { items, staticItems } = listInstance;
 
-  const validSortKey = direction && sortKey && items.some(({ props }) => sortKey in props);
+  const validSortKey = !!direction && !!sortKey && items.some(({ props }) => sortKey in props);
 
   if (!validSortKey) listInstance.restoreItemsOrder();
   else {
@@ -48,11 +50,11 @@ export const sortListItems = async (
 
       if (isDate || isNumber) {
         const [firstItemNumber, secondItemNumber] = [firstItemValue, secondItemValue].map((value) =>
-          isDate ? new Date(value).getTime() : parseFloat(value)
+          isDate ? new Date(value).getTime() : normalizeNumber(value)
         );
 
-        if (isNaN(firstItemNumber)) return 1;
-        if (isNaN(secondItemNumber)) return -1;
+        if (typeof firstItemNumber !== 'number' || isNaN(firstItemNumber)) return 1;
+        if (typeof secondItemNumber !== 'number' || isNaN(secondItemNumber)) return -1;
 
         if (direction === 'asc') return firstItemNumber - secondItemNumber;
 
@@ -69,6 +71,19 @@ export const sortListItems = async (
 
       return secondItemValue.localeCompare(firstItemValue, undefined, collatorOptions);
     });
+
+    // Move static items back to their position
+    // TODO: Remove this once cmscore@1.8.0 has rolled out
+    if (checkCMSCoreVersion('>=', '1.8.0')) {
+      for (const staticItem of staticItems) {
+        const currentIndex = items.indexOf(staticItem);
+
+        if (currentIndex < 0 || typeof staticItem.staticIndex !== 'number') continue;
+
+        items.splice(currentIndex, 1);
+        items.splice(staticItem.staticIndex, 0, staticItem);
+      }
+    }
   }
 
   // Render the new order