+++ packages/cmsnest/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @finsweet/attributes-cmsnest
 
+## 1.7.1
+
+### Patch Changes
+
+- Updated dependencies [4792998a]
+- Updated dependencies [4792998a]
+  - @global/constants@1.2.0
+  - @finsweet/attributes-cmscore@1.8.0
+  - @global/factory@1.1.5
+
+## 1.7.0
+
+### Minor Changes
+
+- 07f32375: Created `fs-cmsattribute` Attribute
+
+### Patch Changes
+
+- Updated dependencies [07f32375]
+  - @global/constants@1.1.0
+  - @global/factory@1.1.4
+  - @finsweet/attributes-cmscore@1.7.2
+
+## 1.6.10
+
+### Patch Changes
+
+- Updated dependencies [d8a4bf6f]
+  - @finsweet/attributes-cmscore@1.7.1
+
+## 1.6.9
+
+### Patch Changes
+
+- Updated dependencies [ef56d454]
+  - @finsweet/attributes-cmscore@1.7.0
+
 ## 1.6.8
 
 ### Patch Changes
@@ -95,11 +132,11 @@ It can be used like:
 <script>
   window.fsAttributes = window.fsAttributes || [];
   window.fsAttributes.push([
-    "cmsnest",
-    cmsLists => {
-      console.log("Attribute has successfully loaded!");
+    'cmsnest',
+    (cmsLists) => {
+      console.log('Attribute has successfully loaded!');
       console.log(cmsLists);
-    }
+    },
   ]);
 </script>
 ```