+++ packages/docs/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @finsweet/attributes-docs
 
+## 1.10.0
+
+### Minor Changes
+
+- bbeca5d2: Added readtime schema
+
+## 1.9.0
+
+### Minor Changes
+
+- 4792998a: Created new `fs-cmsstatic` Attribute.
+
+### Patch Changes
+
+- Updated dependencies [4792998a]
+  - @global/constants@1.2.0
+  - @global/factory@1.1.5
+  - @finsweet/attributes-codehighlight@1.4.1
+  - @finsweet/attributes-copyclip@1.6.1
+  - @finsweet/attributes-toc@1.1.1
+
+## 1.8.0
+
+### Minor Changes
+
+- 07f32375: Created `fs-cmsattribute` Attribute
+
+### Patch Changes
+
+- Updated dependencies [07f32375]
+  - @global/constants@1.1.0
+  - @finsweet/attributes-codehighlight@1.4.0
+  - @finsweet/attributes-copyclip@1.6.0
+  - @finsweet/attributes-toc@1.1.0
+  - @global/factory@1.1.4
+
 ## 1.7.13
 
 ### Patch Changes