+++ packages/favcustom/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @finsweet/attributes-favcustom
 
+## 1.4.1
+
+### Patch Changes
+
+- Updated dependencies [4792998a]
+  - @global/constants@1.2.0
+  - @global/factory@1.1.5
+
+## 1.4.0
+
+### Minor Changes
+
+- 07f32375: Created `fs-cmsattribute` Attribute
+
+### Patch Changes
+
+- Updated dependencies [07f32375]
+  - @global/constants@1.1.0
+  - @global/factory@1.1.4
+
 ## 1.3.5
 
 ### Patch Changes
@@ -53,11 +73,11 @@ It can be used like:
 <script>
   window.fsAttributes = window.fsAttributes || [];
   window.fsAttributes.push([
-    "favcustom",
-    linkHref => {
-      console.log("Attribute has successfully loaded!");
+    'favcustom',
+    (linkHref) => {
+      console.log('Attribute has successfully loaded!');
       console.log(linkHref);
-    }
+    },
   ]);
 </script>
 ```