+++ packages/readtime/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @finsweet/attributes-readtime
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [4792998a]
+  - @global/constants@1.2.0
+  - @global/factory@1.1.5
+
+## 1.1.0
+
+### Minor Changes
+
+- 07f32375: Created `fs-cmsattribute` Attribute
+
+### Patch Changes
+
+- Updated dependencies [07f32375]
+  - @global/constants@1.1.0
+  - @global/factory@1.1.4
+
 ## 1.0.1
 
 ### Patch Changes